Conceptualising on xmlConverterPomeshhenija object
